backend/requirements.txt
asgiref==3.8.1
astroid==3.3.5
attrs==24.2.0
colorama==0.4.6
dill==0.3.9
Django==5.1.2
djangorestframework==3.15.2
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
platformdirs==4.3.6
psycopg2==2.9.10
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.2
